restore FF performance using tested fallback
